rag-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.1
python-dotenv==1.1.1
qdrant-client==1.15.1
sentence-transformers==5.1.0
numpy==2.3.2
scikit-learn==1.7.1
pypdf==6.0.0
python-docx==1.2.0
rapidfuzz==3.13.0
tiktoken==0.11.0
python-multipart==0.0.20
logging-json==0.6.0